requirements.txt
python-fasthtml==0.4.4
requests==2.25.1
ics==0.7.2

